--- packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/comments/jsdoc-component.ts
@@ -1,7 +1,0 @@
-import * as ts from 'typescript';
-import { CommentTags } from '../../../CommentTags';
-import { getNodeJsDocTag } from './jsdoc-helpers';
-
-export function hasUXPinComponentComment(node: ts.Node): boolean {
-  return getNodeJsDocTag(node, CommentTags.UXPIN_COMPONENT) !== undefined;
-}
