+++ packages/uxpin-merge-cli/src/steps/serialization/component/implementation/typescript/serializeTSComponent.ts
@@ -12,9 +12,7 @@ import { ImplSerializationResult } from '../ImplSerializationResult';
 import { PropDefinitionParsingResult } from '../PropDefinitionParsingResult';
 import { PropDefinitionSerializationResult } from '../PropDefinitionSerializationResult';
 import { getComponentDeclaration } from './component/getComponentDeclaration';
-import { getComponentDocUrl } from './comments/jsdoc-docurl';
-import { getComponentNamespace } from './comments/jsdoc-namespace';
-import { getComponentUsePortal } from './comments/jsdoc-useportal';
+import { getComponentDocUrl, getComponentNamespace, getComponentUsePortal } from './comments/jsdoc-uxpin-annotations';
 import { getComponentName } from './component/getComponentName';
 import { getComponentWrappers } from './component/getComponentWrappers';
 import { ComponentDeclaration } from './component/getPropsTypeAndDefaultProps';
